[XNNPACK][Partitioner] SDPA Config
Pull Request resolved: https://github.com/pytorch/executorch/pull/4764

We add the SDPA Config here for partitioner.

Currently there is an issue with SDPA when used from the FairSeq Multihead attention models, so I currently have it disabled for the base partitioner until we resolve that. Otherwise, for our tests, we can use the SDPA correctly from there. We have to track D60553559. Will follow up on this later.
ghstack-source-id: 238566142
@exported-using-ghexport

Differential Revision: [D60323285](https://our.internmc.facebook.com/intern/diff/D60323285/)